feat: switch to module package
BREAKING CHANGE: the package is not of type `module`
