+++ .editorconfig
@@ -1,0 +1,14 @@
+root = true
+
+[*]
+indent_style = tab
+tab_width = 2
+end_of_line = lf
+charset = utf-8
+trim_trailing_whitespace = true
+insert_final_newline = true
+max_line_length = 100
+
+[*.yml]
+indent_style = space
+indent_size = 2
