+++ .eslintignore
@@ -1,0 +1,10 @@
+.idea
+node_modules
+coverage
+CHANGELOG.md
+src/**/*.js
+src/**/*.cjs
+src/**/*.d.ts
+!.eslintrc.cjs
+!.prettierrc.cjs
+!.github
