+++ .github/workflows/ci-cd.yml
@@ -24,28 +24,7 @@ jobs:
 
       - uses: actions/setup-node@v4
         with:
-          node-version: 16
-
-      - uses: bahmutov/npm-install@v1
-        with:
-          useRollingCache: true
-          install-command: yarn --frozen-lockfile
-
-      - name: "Lint"
-        run: yarn lint -f @react-hookz/gha
-
-  build:
-    name: "Build"
-    runs-on: ubuntu-latest
-    steps:
-      - name: "Checkout"
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 16
+          node-version: 20
 
       - uses: bahmutov/npm-install@v1
         with:
@@ -55,13 +34,16 @@ jobs:
       - name: "Build"
         run: yarn build
 
+      - name: "Lint"
+        run: yarn lint -f @react-hookz/gha
+
   dependabot-merge:
     name: "Dependabot automerge"
     runs-on: ubuntu-latest
     permissions:
       pull-requests: write
       contents: write
-    needs: [ "build", "lint" ]
+    needs: [ "lint" ]
     if: github.actor == 'dependabot[bot]' && github.event_name == 'pull_request'
     steps:
       - uses: fastify/github-action-merge-dependabot@v3
@@ -71,7 +53,7 @@ jobs:
   semantic-release:
     name: "Release"
     runs-on: ubuntu-latest
-    needs: [ "build", "lint" ]
+    needs: [ "lint" ]
     if: github.event_name == 'push' || github.event_name == 'workflow_dispatch'
     outputs:
       new-release-published: ${{ steps.release.outputs.new-release-published }}
@@ -83,7 +65,7 @@ jobs:
 
       - uses: actions/setup-node@v4
         with:
-          node-version: 16
+          node-version: 20
 
       - uses: bahmutov/npm-install@v1
         with:
@@ -95,7 +77,7 @@ jobs:
 
       - name: "Release"
         id: "release"
-        uses: codfish/semantic-release-action@v3.0.0
+        uses: codfish/semantic-release-action@v3
         env:
           GITHUB_TOKEN: ${{ secrets.GH_TOKEN }}
           NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
@@ -103,7 +85,7 @@ jobs:
   gen-contributors:
     name: "Generate Contributors List"
     runs-on: ubuntu-latest
-    needs: [ "build", "lint" ]
+    needs: [ "lint" ]
     if: github.event_name == 'push'
     permissions:
       contents: write
