+++ .gitignore
@@ -1,4 +1,7 @@
 /.idea
-/dist
-/node_modules
-yarn-error.log
+node_modules
+/coverage
+/src/**/*.js
+/src/**/*.cjs
+/src/**/*.d.ts
+/src/**/*.d.cts
