+++ .npmignore
@@ -1,0 +1,3 @@
+*
+!/src/**/*.js
+!/src/**/*.d.ts
