--- .prettierrc.js
@@ -1,3 +1,0 @@
-const config = require("@react-hookz/eslint-config/.prettierrc.js");
-
-module.exports = config;
