+++ package.json
@@ -1,73 +1,87 @@
 {
-  "name": "@react-hookz/eslint-formatter-gha",
-  "description": "Eslint formatter that that outputs GitHub Actions annotations.",
-  "version": "1.0.1",
-  "keywords": [
-    "eslint",
-    "formatter",
-    "github",
-    "eslint-formatter",
-    "eslintformatter",
-    "guthub-actions",
-    "guthubactions"
-  ],
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/react-hookz/eslint-formatter-gha.git"
-  },
-  "bugs": {
-    "url": "https://github.com/react-hookz/eslint-formatter-gha/issues"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "/dist"
-  ],
-  "lint-staged": {
-    "*.{js,jsx,ts,tsx,md,mdx}": "eslint --fix"
-  },
-  "release": {
-    "plugins": [
-      "@semantic-release/commit-analyzer",
-      "@semantic-release/release-notes-generator",
-      "@semantic-release/changelog",
-      "@semantic-release/npm",
-      "@semantic-release/git",
-      "@semantic-release/github"
-    ]
-  },
-  "scripts": {
-    "prepare": "husky install",
-    "lint": "eslint ./ --ext ts,js,tsx,jsx,md,mdx",
-    "lint:fix": "yarn lint --fix",
-    "build": "yarn build:cleanup && yarn build:cjs",
-    "build:cleanup": "rimraf ./dist",
-    "build:cjs": "tsc -p ./tsconfig.build.json"
-  },
-  "dependencies": {
-    "@actions/core": "^1.10.1"
-  },
-  "devDependencies": {
-    "@commitlint/cli": "^17.7.2",
-    "@commitlint/config-conventional": "^17.8.0",
-    "@commitlint/cz-commitlint": "^17.7.2",
-    "@react-hookz/eslint-config": "^1.8.0",
-    "@react-hookz/eslint-formatter-gha": "^1.0.0",
-    "@semantic-release/changelog": "^6.0.3",
-    "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^8.1.0",
-    "@types/eslint": "^8.44.4",
-    "@types/node": "^20.8.6",
-    "husky": "^8.0.3",
-    "lint-staged": "^14.0.1",
-    "prettier": "^2.8.8",
-    "rimraf": "^5.0.5",
-    "semantic-release": "^19.0.5",
-    "typescript": "^5.2.2",
-    "yarn": "^1.22.19"
-  }
+	"name": "@react-hookz/eslint-formatter-gha",
+	"description": "Eslint formatter that that outputs GitHub Actions annotations.",
+	"version": "1.0.1",
+	"keywords": [
+		"eslint",
+		"formatter",
+		"github",
+		"eslint-formatter",
+		"eslintformatter",
+		"guthub-actions",
+		"guthubactions"
+	],
+	"license": "MIT",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/react-hookz/eslint-formatter-gha.git"
+	},
+	"bugs": {
+		"url": "https://github.com/react-hookz/eslint-formatter-gha/issues"
+	},
+	"publishConfig": {
+		"access": "public"
+	},
+	"type": "module",
+	"main": "./src/index.js",
+	"types": "./src/index.d.ts",
+	"sideEffects": false,
+	"exports": {
+		".": {
+			"require": "./src/index.cjs"
+		}
+	},
+	"engines": {
+		"node": ">=18.0.0"
+	},
+	"dependencies": {
+		"@actions/core": "^1.10.1"
+	},
+	"devDependencies": {
+		"@commitlint/cli": "^18.6.0",
+		"@commitlint/config-conventional": "^18.6.0",
+		"@commitlint/cz-commitlint": "^18.6.0",
+		"@react-hookz/eslint-config": "^2.1.5",
+		"@react-hookz/eslint-formatter-gha": "^1.0.0",
+		"@semantic-release/changelog": "^6.0.3",
+		"@semantic-release/git": "^10.0.1",
+		"@semantic-release/github": "^9.2.6",
+		"@types/eslint": "^8.56.2",
+		"@types/node": "^20.11.16",
+		"eslint": "^8.56.0",
+		"eslint-plugin-prettier": "^5.1.3",
+		"husky": "^9.0.10",
+		"lint-staged": "^15.2.1",
+		"prettier": "^3.2.5",
+		"rimraf": "^5.0.5",
+		"semantic-release": "^23.0.0",
+		"typescript": "^5.3.3",
+		"yarn": "^1.22.21"
+	},
+	"scripts": {
+		"prepare": "husky install",
+		"commit": "git-cz",
+		"build": "yarn build:clean && tsc --version && tsc",
+		"build:clean": "yarn rimraf -g ./src/**/*.js ./src/**/*.d.ts",
+		"lint": "eslint .",
+		"lint:fix": "yarn lint --fix --fix-type problem,suggestion,layout"
+	},
+	"config": {
+		"commitizen": {
+			"path": "@commitlint/cz-commitlint"
+		}
+	},
+	"lint-staged": {
+		"*.{js,jsx,ts,cts,tsx,md,mdx}": "eslint --fix"
+	},
+	"release": {
+		"plugins": [
+			"@semantic-release/commit-analyzer",
+			"@semantic-release/release-notes-generator",
+			"@semantic-release/changelog",
+			"@semantic-release/npm",
+			"@semantic-release/git",
+			"@semantic-release/github"
+		]
+	}
 }
