--- tsconfig.build.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": [
-    "src/index.ts"
-  ],
-  "compilerOptions": {
-    "noEmit": false,
-    "outDir": "./dist",
-    "declaration": true,
-    "sourceMap": false
-  }
-}
