+++ tsconfig.eslint.json
@@ -1,17 +1,22 @@
 {
-  "extends": "./tsconfig.json",
-  "exclude": [
-    "./dist",
-    "./node_modules"
-  ],
-  "include": [
-    ".**/*.js",
-
-    "**/.*.js",
-    "**/*.js",
-    "**/*.ts",
-    "**/*.tsx",
-    "**/*.md",
-    "**/*.mdx",
-  ]
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"noEmit": true
+	},
+	"exclude": [
+		"./dist",
+		"./node_modules",
+	],
+	"include": [
+		".**/*.js",
+		".**/*.cjs",
+		"**/.*.js",
+		"**/.*.cjs",
+		"**/*.js",
+		"**/*.ts",
+		"**/*.cts",
+		"**/*.tsx",
+		"**/*.md",
+		"**/*.mdx"
+	]
 }
