+++ tsconfig.json
@@ -1,21 +1,25 @@
 {
   "compilerOptions": {
-    "strict": true,
-    "alwaysStrict": true,
-    "target": "ES5",
-    "module": "CommonJS",
-    "moduleResolution": "Node",
     "lib": [
       "ESNext"
     ],
-    "strictNullChecks": true,
-    "strictBindCallApply": true,
-    "noEmit": true,
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "sourceMap": true
+    "target": "ESNext",
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
+    "declaration": true,
+    "removeComments": true,
+    "newLine": "lf",
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "rootDir": "./src"
   },
-  "exclude": [
-    "node_modules"
-  ]
+  "include": [
+    "src/*.ts",
+    "src/*.cts"
+  ],
+  "ts-node": {
+    "esm": true,
+    "swc": true
+  }
 }
